refactor(domain): migration de CurvePoint vers BigDecimal
- Conversion des champs term et value de Double vers BigDecimal
- Définition de la précision DECIMAL(10,4) dans le schéma SQL
- Ajout de la validation @DecimalMin("0.0") sur les champs numériques
- Mise à jour du test CurvePointTests pour utiliser BigDecimal

+++ src/main/resources/database/schema.sql
@@ -56,8 +56,9 @@ CREATE TABLE curve_point (
   curve_point_id INT NOT NULL AUTO_INCREMENT,
   curve_id INT,
   as_of_date TIMESTAMP,
-  term DOUBLE ,
-  value DOUBLE ,
+    -- Maturité (en années fractionnaires) et valeur associée (taux ou rendement)
+    term DECIMAL(10,4),
+    value DECIMAL(10,4),
   creation_date TIMESTAMP ,
 
   PRIMARY KEY (curve_point_id)
